fix context writing
